+++ src/app.module.ts
@@ -13,8 +13,8 @@ import { jobProvider } from './jobs/queue.provider';
 import { RedisService } from './cache/redis.service';
 import { JwtAuthService } from './auth/jwtAuth.service';
 import { JwtAuthGuard } from './auth/auth.guard';
-import { createTransporter } from './mailer/application.mailer';
-import { WelcomeEmail } from './mailer/welcome.email';
+import { createTransporter } from './mailer/mailer.provider';
+import { WelcomeEmail } from './mailer/welcome.mailer';
 
 const axios = require('axios');
