+++ src/mailer/example.ts
@@ -1,5 +1,5 @@
-import { createTransporter, mockTransporter } from './application.mailer';
-import { WelcomeEmail } from './welcome.email';
+import { createTransporter, mockTransporter } from './mailer.provider';
+import { WelcomeEmail } from './welcome.mailer';
 
 createTransporter().then(transporter => {
   WelcomeEmail(transporter)({ to: 'example@example.com', name: 'Miguel' });
