+++ src/users/users.controller.spec.ts
@@ -6,8 +6,8 @@ import { DatabaseModule } from '../databases/database.module';
 import { JobProvider } from '../jobs/application.job';
 import { JwtAuthService } from '../auth/jwtAuth.service';
 import { JwtAuthGuard } from '../auth/auth.guard';
-import { mockTransporter } from '../mailer/application.mailer';
-import { WelcomeEmail } from '../mailer/welcome.email';
+import { mockTransporter } from '../mailer/mailer.provider';
+import { WelcomeEmail } from '../mailer/welcome.mailer';
 
 describe('UsersController', () => {
   let usersController: UsersController;
